Added Custom Exceptions for Controllers

+++ Service/EventService.cs
@@ -1,5 +1,7 @@
 ﻿using AutoMapper;
 using Microsoft.Extensions.Logging;
+using OpenQA.Selenium.DevTools.V108.Page;
+using tajmautAPI.Exceptions;
 using tajmautAPI.Interfaces;
 using tajmautAPI.Interfaces_Service;
 using tajmautAPI.Models;
@@ -28,25 +30,44 @@ namespace tajmautAPI.Service
             {
                 return true;
             }
-            return false;
+            throw new CustomNotFoundException($"Event with ID:{eventId} not found!");
         }
 
         //create event
         public async Task<EventRESPONSE> CreateEvent(EventPostREQUEST request)
         {
-            var getResult = await _repo.CreateEvent(request);
-
-            if(getResult != null)
+            //check if category and restaurant exist
+            if (await _helper.CheckIdRestaurant(request.RestaurantId))
             {
-                var result = await _repo.AddToDB(getResult);
-                return _mapper.Map<EventRESPONSE>(result);
+                if(await _helper.CheckIdCategory(request.CategoryEventId))
+                {
+                    var getResult = await _repo.CreateEvent(request);
+                    if (getResult != null)
+                    {
+                        var result = await _repo.AddToDB(getResult);
+                        return _mapper.Map<EventRESPONSE>(result);
+                    }
+                }
+                else
+                {
+                    throw new CustomBadRequestException("Invalid category ID");
+                }
             }
-            return null;
+            else
+            {
+                throw new CustomBadRequestException("Invalid restaurant ID");
+            }
+
+            throw new CustomBadRequestException("Invalid input");
         }
 
         //delete event by id
         public async Task<EventRESPONSE> DeleteEvent(int eventId)
         {
+            //if invalid input
+            if (eventId < 0)
+                throw new CustomBadRequestException("Invalid ID");
+
             var result = await _repo.DeleteEvent(eventId);
 
             if(result != null)
@@ -54,12 +75,18 @@ namespace tajmautAPI.Service
                 var returnResult = await _repo.DeleteEventDB(result);
                 return _mapper.Map<EventRESPONSE>(returnResult);
             }
-            return null;
+
+            throw new CustomNotFoundException("Event not found!");
+            
         }
 
         //filter events by category
         public async Task<List<EventGetRESPONSE>> FilterEventsByCategory(int categoryId)
         {
+            //invalid input
+            if (categoryId < 0)
+                throw new CustomBadRequestException("Invalid ID");
+
             //check if category exists
             if(await _helper.CheckIdCategory(categoryId))
             {
@@ -70,17 +97,28 @@ namespace tajmautAPI.Service
                     var listEvents = resultEvents.Where(n=>n.CategoryEventId== categoryId).ToList();
                     return await GetEventsWithOtherData(listEvents);
                 }
+                else
+                {
+                    throw new CustomNotFoundException("No data found!");
+                }
             }
-            return null;
+            throw new CustomBadRequestException("Category not found!");
         }
 
         //filter events by city
         public async Task<List<EventGetRESPONSE>> FilterEventsByCity(string city)
         {
+            //invalid input
+            if (city == null)
+                throw new CustomBadRequestException("Invalid input!");
+
             //filter
             var result = await _repo.FilterEventsInCity(city);
-
-            return await GetEventsWithOtherData(result);
+            if (result.Count()>0)
+            {
+                return await GetEventsWithOtherData(result);
+            }
+            throw new CustomNotFoundException("No data found!");
         }
 
         //filter events by date from-to
@@ -88,37 +126,71 @@ namespace tajmautAPI.Service
         {
             var allEvents = await _repo.GetAllEvents();
 
-            //query
-            var sendEvents = allEvents.Where(e=>e.DateTime>=startDate && e.DateTime<=endDate).ToList();
+            //get all events
+            if(allEvents.Count()>0)
+            {
+                //query
+                var sendEvents = allEvents.Where(e => e.DateTime >= startDate && e.DateTime <= endDate).ToList();
+                if(sendEvents.Count()>0)
+                {
+                    return await GetEventsWithOtherData(sendEvents);
+                }
+                throw new CustomNotFoundException("No data found!");
+            }
+            throw new CustomNotFoundException("No data found!");
 
-            return await GetEventsWithOtherData(sendEvents);
         }
 
         //get all events to list
         public async Task<List<EventGetRESPONSE>> GetAllEvents()
         {
             var getResult = await _repo.GetAllEvents();
-            return await GetEventsWithOtherData(getResult);
+
+            if(getResult.Count()>0)
+            {
+                return await GetEventsWithOtherData(getResult);
+            }
+            throw new CustomNotFoundException("No data found!");
         }
 
         //get events from specific restaurant
         public async Task<List<EventGetRESPONSE>> GetAllEventsByRestaurant(int restaurantId)
         {
+            //invalid input
+            if (restaurantId < 0)
+                throw new CustomBadRequestException("Invalid ID");
+
             var allEvents = await _repo.GetAllEvents();
+            if (allEvents.Count() > 0)
+            {
+                //query
+                var listEvents = allEvents.Where(n => n.RestaurantId == restaurantId).ToList();
+                if(listEvents.Count()>0)
+                {
+                    return await GetEventsWithOtherData(listEvents);
+                }
+                throw new CustomNotFoundException("No data found!");
+            }
+            throw new CustomNotFoundException("No data found!");
 
-            //query
-            var listEvents = allEvents.Where(n=>n.RestaurantId== restaurantId).ToList();
-
-            return await GetEventsWithOtherData(listEvents);
         }
 
         //get event by id
         public async Task<List<EventGetRESPONSE>> GetEventById(int eventId)
         {
+
+            //invalid input
+            if (eventId < 0)
+                throw new CustomBadRequestException("Invalid ID");
+
             var result = await _repo.GetEventById(eventId);
 
-            return await GetEventsWithOtherData(result);
+            if(result.Count() > 0)
+            {
+                return await GetEventsWithOtherData(result);
+            }
 
+            throw new CustomNotFoundException("No data found!");
         }
 
         //get other data for the events in sorted list
@@ -179,18 +251,35 @@ namespace tajmautAPI.Service
         //update event
         public async Task<EventRESPONSE> UpdateEvent(EventPostREQUEST request, int eventId)
         {
+
+            //invalid input
+            if (eventId < 0)
+                throw new CustomBadRequestException("Invalid ID");
+
             var resultEvent = await _repo.UpdateEvent(request, eventId);
 
+            //if found
             if (resultEvent != null)
             {
                 //check if restaurant and category exist
-                if (await _helper.CheckIdCategory(request.CategoryEventId) && await _helper.CheckIdRestaurant(request.RestaurantId))
+                if (await _helper.CheckIdCategory(request.CategoryEventId))
                 {
-                    var result = await _repo.SaveUpdatesEventDB(resultEvent, request);
-                    return _mapper.Map<EventRESPONSE>(result);
+                    if(await _helper.CheckIdRestaurant(request.RestaurantId))
+                    {
+                        var result = await _repo.SaveUpdatesEventDB(resultEvent, request);
+                        return _mapper.Map<EventRESPONSE>(result);
+                    }
+                    else
+                    {
+                        throw new CustomBadRequestException("Invalid Restaurant ID");
+                    }
+                }
+                else
+                {
+                    throw new CustomBadRequestException("Invalid Category ID");
                 }
             }
-            return null;
+            throw new CustomNotFoundException("No data found");
         }
     }
 }
